requirements.txt
pycord[voice]
python-dotenv
pymongo
